requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.1
typer[all]>=0.12.3
rich>=13.7.1
streamlit>=1.35.0
httpx>=0.27.0
pydantic>=2.7.1
